Moved into Utils for utility.
